WebUI work, repaired Master-Slave temp diffs

+++ Labs/Lab4_2/CommProtocol.cpp
@@ -73,7 +73,7 @@ void CommProtocol_::SendMessage(TempMessage message) {
 
     for(int i = 1; i < 5; i++) {
         writing_msg[2 * i] = (byte)(temps[i - 1] + NUMBER_PADDING);
-        writing_msg[2 * i + 1] = (byte)((byte)(temps[i - 1] * 100) % 100 + NUMBER_PADDING);
+        writing_msg[2 * i + 1] = (byte)((int)(temps[i - 1] * 100) % 100 + NUMBER_PADDING);
     }
 
     // Write start token
